3348-smallest-divisible-digit-product-ii

+++ 0-common-lib/src/gcd.rs
@@ -1,3 +1,3 @@
-pub fn gcd(a: usize, b: usize) -> usize {
+pub fn gcd(a: u64, b: u64) -> u64 {
     if a == 0 { b } else { gcd(b % a, a) }
 }
